Project 2 the First Part Ready
The two page classes (directory_page and bucket_page) are implemented
and well tested. Granted that, some of the APIs are still needed to
tweak when trying to put them into the project part 2 (extendible hash
table). On top of that, concurrency control has not been considered up
to the point.